[REFACTOR] introduce MotherPlantEntityTable
